abstracted mass container to separate component
