+++ src/ServerContainer.jsx
@@ -1,6 +1,6 @@
 export default function ServerContainer({ name, index, clickFunction }) {
     return (
-        <div className="border-black border-x border-y ml-1"
+        <div className="border-black border-x border-y"
             key={index}
             accessKey={index}
             onClick={clickFunction}
